requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
uvicorn[standard]==0.30.6
pydantic>=2.7.0,<3.0.0
pandas==2.2.2
kafka-python==2.0.2
redis==5.0.8